+++ CHANGES.md
@@ -9,6 +9,7 @@ Changes in 11.1
 - Set email for default admin user so he can reset his password in InstallDatabase.php
 - Prevent using App name, username, or email in the password in PasswordReset.php, FirstLogin.fnc.php, Inputs.php & Preferences.php
 - JS Add userInputs param to prevent using App name, username, or email in the password in jquery-passwordstrength.js
+- Return EMAIL column for students too (empty if "Student email field" not set) in User.fnc.php
 
 Changes in 11.0.2
 -----------------